use bigger epsilon

+++ ac3d.h
@@ -348,7 +348,7 @@ private:
         }
         bool equals(const Point3 &other) const
         {
-            static constexpr double  SMALL_NUM = static_cast<double>(std::numeric_limits<double>::epsilon());
+            static constexpr double  SMALL_NUM = static_cast<double>(std::numeric_limits<float>::epsilon());
 
             return std::abs(x() - other.x()) < SMALL_NUM &&
                    std::abs(y() - other.y()) < SMALL_NUM &&
